future::then: use std::terminate instead of abort
If an exception tries to pass a noexcept function, std::terminate is
called instead.

In practice this is the same as calling abort, but the call to
std::terminate is done by the unwind runtime, so the it only shows
up as metadata in the exception handling table.

The release size of scylla goes from

   text	   data	    bss	    dec	    hex	filename
82541296	  35824	 275984	82853104	4f03cf0	build/release/scylla

to

   text	   data	    bss	    dec	    hex	filename
81228156	  35824	 275984	81539964	4dc337c	build/release/scylla

Message-Id: <20190313163926.74589-1-espindola@scylladb.com>