reactor: improve detection of io_pgetevents()
Some versions of docker have a syscall whitelist, which is naturally out-of-date
and does not include io_pgetevents(). While our detection code calls the wrapper,
the wrapper avoids the syscall when possible, which is the case if IOCB_CMD_POLL
works but io_pgetevents() is blocked.

Improve the detection by passing true for the force_syscall parameter, which forces
io_pgetevents() to be called. If docker blocks the syscall, then we'll disable the
aio backend for this run.
Message-Id: <20190318183749.19889-1-avi@scylladb.com>