tests: Show a test's output when it fails
Previously, no output from a test would be printed by default by
`test.py` (since this is the behavior of CTest).

It's possible to show more information by running `test.py` or CTest
with `--verbose`, but this shows output for all tests irrespective of
whether or not they were successful.

With this change, we run CTest with `--output-on-failure`, which will
not show a test's output unless it fails.

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <3a78da8d96c74e9602871747d47fcb0d30b4aaba.1553003045.git.jhaberku@scylladb.com>