build: Allow cooked dependencies with configure.py
The `configure.py` script is quite convenient, but depends on the
availability of dependencies on the system.

`cooking.sh` can be used to fetch dependencies locally for development,
but then the convenient option processing done by `configure.py` has to
be reproduced by hand.

This change adds the (repeatable) `--cook $dependency_name` option to
`configure.py`, where `$dependency_name` is an ingredient described in
`cooking_recipe.cmake`.

This way, for example, Seastar can be configured with a local version of
fmt as follows:

    ./configure.py --mode=dev --cook fmt

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <89b514c299102b7012bf064467d23d758dba15fc.1553024861.git.jhaberku@scylladb.com>