Allow when_all and when_all_succeed to take function arguments
The issue with the existing interface is that it is easy to create
some of the futures that one wants to wait for and then hit an
exception. For example, in

when_all(create_fut(), create_fut());

one call to create_fut might return a future, but the next one can
throw.

By passing functions that return futures to when_all and
when_all_succeed, we can convert exceptions into failed futures.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190321212715.1386-3-espindola@scylladb.com>