test: verify that negative sleep time doesn't cause infinite sleep
When debugging a Seastar application a worry was raised that perhaps
a negative sleep_abortable() does not finish immediately, and maybe
because of some bug sleeps forever. The theory was that maybe we have
a bug in our timer implementation, where timers set for the past will
never expire.

So I wrote this test to verify that sleep_abortable(-10s) in fact does
not hang, and there is apparently no such bug.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190402103308.21512-1-nyh@scylladb.com>