seastar-addr2line: print prefix of line containing the first address
In some cases the information which helps identify a certain backtrace
is found on the line of the first address of the backtrace, before the
address, e.g.:
    Apr  1 00:52:29 scdb1 scylla:  [shard 12] seastar_memory - oversized allocation: 5169152 bytes, please report: at   0x7cda88

In this case, printing the lines before the trace are not helpful, one
has to see the prefix of the line containing the first address in order
to identify the backtrace. To fix this, in addition to a
user-configurable amount of lines before the backtrace, also print the
prefix of the line of the first address before the resolved backtrace is
printed.