seastar-addr2line: add --match command-line option
Allow optionally filtering backtraces and only resolving those that
match the filter. Helpful when one is resolving all traces from a huge
log file but is interested only in certain type of backtraces.
The filtering is done on the context of the backtrace: the lines before
it (their number is configurable by --before) and the prefix of the
line of the first address in the backtrace (see previous patch).