core/memory: Make small pools use the full span size
The span size we get from the large allocator can be larger than the
requested size due to internal constraints of the buddy allocator.

The small allocator won't use the extra pages in the span. This is a
waste.  It also corrupts _pages_in_use, which is incremented with the
requested size, but decremented using allocated span size when it's
freed.

Fix by using the full span.

Fixes #625
Fixes #608

Message-Id: <1554298641-28804-1-git-send-email-tgrabiec@scylladb.com>