doc/tutorial.md: reword mention of seastar::thread premption on get()
The tutorial's wording suggesting that get() calls are blocking "points"
where a seastar::thread can be preempted. Naming such locations a "point"
may suggest that even when get() does not in fact block, preemption can
happen at that point. And this was indeed the case until commit
7e2483ac867d7deff6fa6170c86667e1f5c147b0. But after this commit, preemption
can only happen if the get() call actually blocks. So change the wording
to be clearer on that.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190404095022.5921-1-nyh@scylladb.com>