reactor: make metric about task quota violations less sensitive
We have noticed in the field that this metric can be unfortunately a bit
misleading. It accumulates whatever time we ran for more than the task
quota, even if we only missed the quota for a couple of microseconds.
Those are not a big problem and are just a fact of life since the timer
thread won't fire at the exact time, continuations won't check
need_preempt continuously, etc.

We saw in the field that as a result of that, we can see this metric
showing as moderately high, even though there are no backtraces in the
logs whatsoever even at low rates.

This patch changes our threshold for accumulation and only accumulates
the time spent in violation if we violated it by at least an extra task
quota.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190411140329.16837-1-glauber@scylladb.com>