future::then_wrapped: use std::terminate instead of abort
If an exception tries to pass a noexcept function, std::terminate is
called instead.

In practice this is the same as calling abort, but the call to
td::terminate is done by the unwind runtime, so the it only shows
up as metadata in the exception handling table.

The release size of scylla goes from

text     data  bss    dec      hex     filename
81077472 40896 271016 81389384 4d9e748 build/release/scylla

to

text     data  bss    dec      hex     filename
79840020 40896 271016 80151932 4c7057c build/release/scylla

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190412194710.110362-1-espindola@scylladb.com>