tests: futures_test: allow project to compile with C++14
When compiling the project with C++14

    $ ./configure.py --mode=dev --cook=fmt --c++-dialect=c++14

You get

    [141/204] Building CXX object tests/unit/CMakeFiles/test_unit_futures.dir/futures_test.cc.o
    FAILED: tests/unit/CMakeFiles/test_unit_futures.dir/futures_test.cc.o
    /usr/bin/g++  -DBOOST_TEST_DYN_LINK -DSEASTAR_HAS_MEMBARRIER -DSEASTAR_HAVE_ASAN_FIBER_SUPPORT -DSEASTAR_HAVE_GCC6_CONCEPTS -DSEASTAR_HAVE_HWLOC -DSEASTAR_HAVE_LZ4_COMPRESS_DEFAULT -DSEASTAR_HAVE_NUMA -DSEASTAR_TESTING_MAIN -DSEASTAR_TYPE_ERASE_MORE -I../../tests/unit -I../../src -I../../include -Igen/include -isystem _cooking/installed/include -O1   -std=c++14 -U_FORTIFY_SOURCE -fconcepts -fvisibility=hidden -UNDEBUG -Wall -Werror -Wno-error=deprecated-declarations -gz -MD -MT tests/unit/CMakeFiles/test_unit_futures.dir/futures_test.cc.o -MF tests/unit/CMakeFiles/test_unit_futures.dir/futures_test.cc.o.d -o tests/unit/CMakeFiles/test_unit_futures.dir/futures_test.cc.o -c ../../tests/unit/futures_test.cc
    ../../tests/unit/futures_test.cc: In member function ‘virtual seastar::future<> test_when_all_functions::run_test_case()’:
    ../../tests/unit/futures_test.cc:749:14: error: use of deleted function ‘non_copy_non_move::non_copy_non_move(non_copy_non_move&&)’
         auto f = [x = non_copy_non_move{}] {
                  ^
    ../../tests/unit/futures_test.cc:744:5: note: declared here
         non_copy_non_move(non_copy_non_move&&) = delete;
         ^~~~~~~~~~~~~~~~~
    ../../tests/unit/futures_test.cc: In member function ‘virtual seastar::future<> test_when_all_succeed_functions::run_test_case()’:
    ../../tests/unit/futures_test.cc:770:14: error: use of deleted function ‘non_copy_non_move::non_copy_non_move(non_copy_non_move&&)’
         auto f = [x = non_copy_non_move{}] {
                  ^
    ../../tests/unit/futures_test.cc:744:5: note: declared here
         non_copy_non_move(non_copy_non_move&&) = delete;

I assume that C++17 captures rvalue without moving it, and C++14 moves
the rvalue. I'm far from a C++ expert but from reading the C++14
standard http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2013/n3690.pdf
§5.2.1 page 92 line 11:

For every init-capture a non-static data member named by the identifier of the init-capture is declared in
the closure type. This member is not a bit-field and not mutable. The type of that member corresponds to
the type of a hypothetical variable declaration of the form “auto init-capture ;”, except that the variable
name (i.e., the identifier of the init-capture) is replaced by a unique identifier.
[ Note: This enables an initcapture like “x = std::move(x)”; the second “x” must bind to a declaration in
the surrounding context. — end note ] No entity is captured by an init-capture.
Within the lambda-expressions lambda-declarator and
compound-statement, the identifier in the init-capture hides any declaration of the same name in scopes
enclosing the lambda-expression.

I understand there's no special requirement for rvalues in init
captures.

I conclude the C++14 can do whatever it pleases with rvalue in
init-capture, and therefor the proper solution here is to remove this
test when compiling with C++14.

Signed-off-by: Elazar Leibovich <elazarl@gmail.com>
Message-Id: <20190415131051.21443-2-elazarl@gmail.com>