app_template: allow opting out of automatic SIGINT/SIGTERM handling
The automatic SIGINT/SIGTERM handling is veru coarse and requires careful
registration of atexit tasks. Allow opting out of it for new (or updated)
applications.

When opted out, there is no special handling of SIGINT/SIGTERM - the
default signal disposition (terminate) is kept. The user can override
this with reactor::handle_signal(), and if desired use that to terminate
the program by returning from the callback given to app_template::run().
If that callback runs a thread, it is easy to set up RAII teardown of
resources and services.

Later, we should consider switching the default.

Ref #261.
Message-Id: <20190418085452.28283-1-avi@scylladb.com>