build: Restore DPDK machine inference from cflags
Previously, 3b00827d07cbcad09a4e6d0cbb3b459f825b600e inferred the
"machine" parameter for DPDK based on the presence of `-march` in
Seastar's cflags (if it was present).

This ensured that DPDK would target the same architecture as Seastar.

This functionality was omitted in the transition to CMake, and this
change restores it.

We define a new CMake cache variable -- `Seastar_DPDK_MACHINE` -- which
defaults to "native" and is forwarded to the configuration of DPDK. The
value of this variable overrides any value in `dpdk_config`.

This value can be set during the invocation of `cmake` or of
`cooking.sh`.

For convenience, `configure.py` will automatically set the value of
`Seastar_DPDK_MACHINE` based on the value of `--cflags` (as it did
before the switch to CMake).

When Seastar is configured for DPDK support, we need the selected DPDK
architecture to be transitively applied to consumers of Seastar as well.

Previously, this was accomplished via `Finddpdk.cmake`, but now this
logic is in `CMakeLists.txt` where it belongs.

If DPDK is enabled but `Seastar_DPDK_MACHINE` is not modified from the
default of "native", then Seastar will be compiled with `-march=native`.

Fixes #632

Tests: unit (dev), dist (dev)
Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <be0a2d8486eabf699a2c796e4a2fba3ad6b863ae.1555600899.git.jhaberku@scylladb.com>