future: change to always have local state
With this change a future always holds its state locally. When there
is an associated promise, it is the promise that follows a pointer
to the future.

This has the advantage that promise already required a _state
pointer. With this patch only promise has to keep track on which
future_state is currently in use.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>