shared_future: avoid creating broken promises
Before this patch the constructor of shared_future calls
future::then_wrapped.

The net result is that just constructing and destructing a
shared_future will cause a broken future exception if the original
future was not available.

Just creating a broken future exception doesn't cause any harm, but it
does make it harder to track down unexpected ones.

With this patch then_wrapped is only called in get_future, which is as
late as we can without returning a proxy.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>