sstring: silence technically correct but unhelpful warning in sstring move ctor
basic_sstring's move constuctor copies the entire (typically 16 byte)
object; but it's possible that only some prefix of the embedded string
was initialized (sstring("hello") only initializes the first 8 bytes).
gcc 9 is able to detect this, and warns. While we indeed copy
uninitialized memory, this is not a problem since the unintitialized
data is not exposed by public methods, without a user bug.

Silence the unhelpful warning with a pragmatic pragma.
Message-Id: <20190504165840.7140-1-avi@scylladb.com>