http: fix request content for POST requests
Commit f43b7ead2c27bf6fe43c381ea5078bb14dfbdf29 added the ability to
retrieve the body of a request, which is important for POST, but did it
incorrectly. This patch reverts the above fix and adds a better fix.

What the original commit did was to look at the network buffer after the
headers ended, and consider what we had there to be the request's content.
However, this was wrong on two levels: First, this was only one network
buffer, and with long request bodies we may need to read additional buffers
for the full content. Second, while content was copied from the network
buffer, the buffer's pointer was *not* advanced, so if we later have another
request on the same connection the parser will continue to parse (and fail)
where the previous request's body began, instead of where it ended.

Instead, our code needs to first parse the headers as it did before, and
only when that is done, read (and really read) from the network input
stream the full body - which may involve multiple network buffers.
The size of the request's body is ought to be specified with a
"Content-Length:" header (there is also the possibility of chunked-
encoding, but this patch doesn't support it yet and it is left as a FIXME).

Note that this code reads the entire request body into a string
request::_content. This is obviously a bad idea - but is the current
interface our request object has, and this patch doesn't change this
interface. In the future, we should probably replace this interface with
an input stream, which the handler could read from to read the entire request
body.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190512093436.31249-1-nyh@scylladb.com>