improve error message on AIO setup failure
Right now we try to set up AIO and assert that we succeeded.  The
assertion is justified since there isn't a way we can proceed if that's
the case.

However, this leaves the user trying to figure out what happened.
Failures to setup AIO are fairly common if one is not running scylla
through a packaged installation: we will fail with the default values in
/proc/sys/fs/aio-max-nr. Our packaging will take care of that but if the
user is going through some other way they will have to increase that
manually.

We should let them know in an error message that this is the most common
cause so we can quickly fix it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190513192531.10956-1-glauber@scylladb.com>