build: Help Seastar to find Boost on Fedora 30
Previously, we unconditionally used `cmake/FindBoost.cmake` when
checking for the availability of Boost on the system during the
configuration process.

This file is vendored from CMake directly, since `FindBoost.cmake`
installed alongside CMake 3.5 does not have awareness of Boost
1.64 (which is the minimum version that we support).

However, on systems with a version of Boost installed which is greater
than the maximum supported by our vendored `FindBoost.cmake`, the
configuration process will print warnings like this:

    CMake Warning at cmake/FindBoost.cmake:839 (message):
     New Boost version may have incorrect or missing dependencies and imported
     targets

The solution is to conditionally use the vendored find-module.

We first attempt to use the CMake-bundled `FindBoost.cmake`. If this
fails (because there is none, or the version of Boost we need is not
supported), then we try again with the vendored `FindBoost.cmake`.

I tested this change with CMake 3.5 and CMake 3.14 on Fedora 30 and
Fedora 29.

Fixes #639

Signed-off-by: Jesse Haber-Kucharsky <jhaberku@scylladb.com>
Message-Id: <6a7d84da304828f705f31ef6cb2487c791c3c757.1557943749.git.jhaberku@scylladb.com>