future: Simplify promise::set_value
The variadic version now just forwards to the std::tuple one.

For future_state we just remove the std::tuple overload since the set
template parameter can be a std::tuple&& or a std::tuple&.

This also fixes the noexcept specification of the variadic version. It
was marked noexcept, but that is not always the case. This is a small
inline function, so it is probably not worth the metaprogramming
necessary to say when exactly this function would be noexcept. The
corresponding future_state::set is also not noexcept.

The added tests check that set_value can still be used with tuples and
that now types that throw in the copy constructor can also be used.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>