core/deleter: Fix abort when append() is called twice with a shared deleter
Appending a deleter which has the same impl pointer as a deleter which
already exists in the chain triggered the assertion which was there to
protect against formation of a cycle in the deleter chain (and thus
infinite loop).

This case can be handled by simply dropping such deleters based on the
premise that the existing deleter is enough to keep the object alive.

This causes the following code to fail:

  temporary_buffer<char> big_buf(200);

  auto buf1 = big_buf.share(0, 10);
  auto buf2 = big_buf.share(11, 10);

  net::packet p;
  p.append(net::packet(std::move(buf1)));
  p.append(net::packet(std::move(buf2)));

Fixes #646.

Tests:

  - unit (dev)

Message-Id: <1559585573-17364-1-git-send-email-tgrabiec@scylladb.com>