Add a Sanitize build mode
For now this is just Debug + -Os, but it will probably become more
specialized over time.

The objective of this mode is to be able to run tests with sanitizers
faster and, more importantly, be able to run sanitized scylla in more
realistic use cases.

On my machine ctest takes 3:48.21 total on a Debug build and 2:51.41
on a Sanitize build.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190605164011.25886-1-espindola@scylladb.com>