net: Fix dereferencing moved object
Seastar crashes when running on Arm with native network stack enabled.

It's caused by a code snippet like "func(free(ptr), use(ptr))", which
is a function call with two parameters. First parameter is a function
to free a ptr, while the second parameter dereferences the ptr.

This code works okay on x86 as parameters are passed from right to left
in stack pushing order. So use(ptr) is called before free(ptr).

But it fails on Arm as parameters are passed from left to right on Arm
to leverage abundant general purpose registers. So free(ptr) will be
called before use(ptr), which leads to crash.

To show the issue, below code outputs "ba" on x86, but "ab" on Arm.

printf("", printf("a"), printf("b"));

Message-Id: <20190611025611.24192-1-yibo.cai@arm.com>