reactor: provide an uptime function
Applications may want to know for how long they have been running.  It
is possible to aquire this information from /proc, but since we already
record start time we can expose this faster using a simpler operation.

There will be differences of uptime in each shard, but this seems like
a reasonable restriction for seastar applications.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190618143807.8728-1-glauber@scylladb.com>