perftune: make sure that ethtool -K has a chance of succeeding
Right now, we have the following code that tries to enable ntuple
offloading

        try:
            run_one_command(['ethtool','-K', iface, 'ntuple', 'on'], stderr=subprocess.DEVNULL)
            print("ok")
        except:
            print("not supported")

This always fails, and we've been essentially ignoring believing that this
is just the way the NIC is.

The reason for the failure is more subtle:

we pass an 'stderr' argument to run_one_command, while the function
expects 'my_stderr' in its signature. If we take this code out of the
try-catch block, here's what we see:

Trying to enable ntuple filtering HW offload for enp30s0...ERROR: run_one_command() got an unexpected keyword argument 'stderr'. Your system can't be tuned until the issue is fixed.

Instead of the usual sneaky:

Trying to enable ntuple filtering HW offload for enp30s0...not supported

This patch fixes the keyword so this won't happen
(also maybe we shouldn't be using a catch all here)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20190621174750.12120-1-glauber@scylladb.com>