scheduling: Add rename functionality to scheduling groups
This patch add the ability to rename a scheduling group.
One reason why one would like to rename a scheduling group
is because destroying them is too strict, the user has to
make sure they are not used anymore and there are no pending
tasks, the later is hard to determine. The renaming affects
the statitstics, this means that when renaming the counters
will be labled under the new name.

A unit test have been added that stress tests the renaming
functionality.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20190716141943.10497-2-eliransin@scylladb.com>