io queues: Add renaming functionality to io priority class
This change adds the support for io priority class renaming.
Since io priority classes cannot be destroyed, and even if
they could, it would probablly would be complicated to tear
down. It can be easier to just reuse the io priority class
for other purpose. In order to facilitate this we support
renaming.

Some special attention has to be given to the possiblity
of rerenaming, it can happen since we want to avoid
holding a lock while making a cross shard operation.

A tests has been added to stress test the renaminc and make
sure it will not generate duplicate statistics exceptions.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20190716141943.10497-3-eliransin@scylladb.com>