Assert that append_challenged_posix_file_impl are closed.
If a append_challenged_posix_file_impl is not closed, we risk having
tasks still running and will try to use freed memory.

With this patch we expand the old _done flag to an enum to track the
stage the file is in and asserts in the destructor that the file has
been fully closed.

The destructor cannot do anything fancier than asserting since it
cannot propagate a future and cannot safely throw an exception.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190721173722.13891-1-espindola@scylladb.com>