execution_stage: prevent unbounded growth
An execution_stage queues function calls for later execution in a batch,
with the intent that batch execution makes better use of the processor
caches and branch predicion. But this generates an assymetry: all the queued
functions are bundled into one task, and if there are many tasks in the
scheduler's task queue, it will have to compete with them for processor time.
Because of the round-robin nature of the scheduler a large number of small
tasks will dominate over a small number of long-running, but preempt-friendly
tasks. This can lead to unbounded queue growth and thus running out of memory.

To prevent this, introduce a maximum queue length. Whenever we reach that we
will flush the queue, freeing at least one slot (since flushing will be limited
by preemption). This prevents the queue from growing without bounds, and restores
some fairness to the system by making the enqueuers pay for dequeuing.

The queue limit was chosen at 1024. This can result in too-early flushing if the
tasks execute in less than half a microsecond (so we won't even reach a preemption
point), but in practice functions in execution stages are much more expensive. If
it becomes a problem, we can make both flush_threshold and max_queue_length
configurable.

Tests: execution_stage_test (dev)
Message-Id: <20190722074344.27040-1-avi@scylladb.com>