reactor: improve handling of connect storms
Currently, accept() (in the posix network stack) does not speculate that a after
a successful accept we'll still have a connection in the listen queue. This means
that to get the next connection, we'll have to poll again.

This makes sense for servers with long-running connections because connects are rare,
and therefore we expect that the listen queue will usually have zero or one connections
waiting. If we attempt to accept() again, we'll just get EAGAIN and waste a system call.

However, there are some scenarios where we do expect a connect storm. One is a benchmark
that uses a lot of connections. Another is a server that is part of a cluster, resuming
after a shutdown. In that case all of the other nodes in the cluster, as well as clients,
will start connecting to it.

In the balance, an occasional wasted syscall in a server that has long-running connections
suffers only a small penalty compared to the penalty we suffer from not speculating
if we have to endure a connect storm, so switch the code to speculate that a connection
will be available after accept(). We also add posix::file_desc::try_accept() in order
to be able to handle an EAGAIN without an exception.

Tests: manual testing with httpd and strace.
Message-Id: <20190725102929.17508-1-avi@scylladb.com>