thread: reduce allocations during context switch
Currently, we allocate a continuation when we switch back to a thread (implicitly
via make_task()). Since tasks can be reused, we can allocate the task just once
(and moreover, allocate it as part of the thread_context) and use it again and again.
This should make context switches a little less expensive.

Since a task object already contains a scheduling_group, we can drop the _attr and
_scheduling_group members which duplicate it.
Message-Id: <20190727164131.24707-1-avi@scylladb.com>