perftune.py: introduce a --get-cpu-mask-quiet parameter
When this new parameter is given perftune.py is going to print the
compute CPU set similarly to --get-cpu-mask but it would also print a
zero CPU set if that's what it turns out to be.
Exit status is going to be zero if the execution is successful apart
from that.

--get-cpu-mask would still print an error and return a non-zero exit
status if a compute CPU set turns out to be zero.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>