core: Add ability to abort on EBADF and ENOTSOCK
Those are typically symptoms of use-after-free or memory corruption in
the program. It's better to catch such error sooner than later.

That situation is also dangerous since if a valid descriptor would
land under the invalid access, not the one which was intended for the
operation, then the operation may be performed on the wrong file and
result in corruption.

This patch adds ability for a seastar program to enable aborts when
those errors happen. By default aborts are disabled.

Message-Id: <1565206020-28899-1-git-send-email-tgrabiec@scylladb.com>