rpc: correctly handle huge timeouts
It is generally a bad idea to pass huge timeouts to RPC, but nevertheless,
any legal timeout given should be honored correctly. It turns out that
we had a bugs when timeout was time_point::max() or duration::max().
The symptom observed before this patch were that negative timeouts were
calculated by the receiver, causing its reply to be dropped by
connection::send() and the sender never receiving a reply.

The bug with duration::max() as timeout was straightforward: When this
relative timeout was converted to an absolute timeout, the sum
now+timeout overflowed and became negative. The straightforward solution
in this patch is to cap the sum at time_point::max().

The cause of the bug with time_point::max() as timeout is more delicate:
Because clocks on different nodes aren't perfectly synchronized, the
sending node converts the timeout from an absolute time to a relative
time (how much time is left), and then the receiving node converts it
back to an absolute time using its own clock. Due to small clock
differences, and to the time it took to send the message, the resulting
absolute time may be slightly larger than the one in the sender. If it
was already the maximal integer (time_point::max()) in the sender,
the calculation will overflow in the receiver and it will calculate
a timeout in the distance past, instead of the distant future as expected!
Again the solution of this patch is to cap the result of the sum at
time_point::max().

This patch also includes unit tests for both cases. Both tests fail
before this patch, and succeed after it. The test for time_point::max()
requires a loop because the overflow doesn't happen every time - it only
happens if the calculation in the sender and receiver happen across a
millisecond boundary. So we need to loop over many milliseconds to
increase the likelyhood of catching this condition. In my experience,
200 milliseconds was more than enough to catch this problem every time.
Note that when these tests do fail, their failure mode is a hung send,
so both tests include extra code to give up after a long enough (3s)
hang, and finish finish the test with a normal failure.

Fixes #671

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190818132502.9816-1-nyh@scylladb.com>