reactor: implement aio fsync
Linux 4.18 and later implement IOCB_CMD_FSYNC, which fdatasync()s a file
via the aio interface instead of a normal syscall interface. This results
in lower latency, as we don't have to queue the fdatasync() syscall in
the syscall thread (which may be doing other things), and fsyncs can be
executed in parallel.

We don't go through the normal io_queue and I/O scheduler mechanism. Most
of the work of an fdatasync() applies to all files that currently have
volatile data: flushing the filesystem log, and flushing the disk write
cache (in case it is not write through). We also can't estimate how much
work that is. So we queue it outside the I/O scheduler.

Tests: observe via strace and --force-aio-syscalls 1.