net: do not retry connection attempt if a requested port cannot be bound
posix_socket_impl:connect() tries to find a specific port (a function
of a shard number the connect() is run on) to bind to if a caller did
not provide a port by itself. But current code will run in a loop if
the port is provided and it is busy. The same will happen if no port was
provided, but there is no any free port to bind to.