net: retry connection attempt on EADDRNOTAVAIL
It turns out that connect() can return EADDRNOTAVAIL if bind() was
called before it and the chosen local address cannot be used (connect()
and bind() use different algorithms to choose the address). We need to
try another port on this error as well. But since the bind() call
succeeded the socket is now bound so the next bind() call will fail
with EINVAL. To solve that move a socket creation into the loop.