shared_future: don't warn on broken futures
Before 27db48ece a shared_future would not itself warn if broken. This
patch reverts that unintended change.

The patch also adds a testcase where we were producing a warning and
there was nothing the user could do since the shared_future is a
private member of the shared_promise.

Fixes #683

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190918151939.24615-1-espindola@scylladb.com>