future: Warn on promise::set_exception with no corresponding future or task
In older versions of seastar a future would access its state via a
pointer to promise. Once the future was destroyed, the promise still
had a valid state and set_exception would record an exception that
would be reported by the promise destructor.

Since 8aa3051e66a833c67d7600cc6e431a20f4782a1a a future always has a
local state, and after it is destroyed without creating a task the
promise has nowhere to store the exception.

Since the only observable difference is the warning, we now issue the
warning in promise::set_exception.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20190918212125.104071-1-espindola@scylladb.com>