rpc: introduce rpc::tuple as a way to move away from variadic future
rpc now supports returning a variadic future from the remote procedure.
Since this is now deprecated, this patch introduces rpc::tuple (a
strong typedef of std::tuple) to encourage users to return a
future<rpc::tuple<a, b>> instead of the deprecated future<a, b>.

I chose to introduce rpc::tuple instead of using std::tuple, since
there may be existing users of std::tuple in rpc, with its own
serialization. Using rpc::tuple means we have complete control of
how it is serialized, and the implementation serializes it in the
same way as multiple return values in a variadic future are, for
compatibilty.

The std::tuple_size and std::tuple_element extension points are
implemented for rpc::tuple so that structured binding works, and a
deduction guide is also supplied for general niceness.
Message-Id: <20190921164944.30783-1-avi@scylladb.com>