build: Include dpdk as a single object in libseastar.a
We always build our own dpdk and always link it statically. That
combined with the fact that dpdk requires --whole-archive to link in
the drivers, makes seastar hard to use: one has to track all
dependency paths that lead to seastar and only link dpdk once.

With this patch we turn dpdk into a single .o file in
libseastar.a. With this change seastar with dpdk enabled can be linked
multiple times without causing duplicated symbol errors.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20191003190904.41319-1-espindola@scylladb.com>