prometheus: Add an option to filter by label
The prometheus API already support filter metrics by their name.

This patch adds an option to filter also by label.

For example, to return all the metrics runs on shard 0, you can use:
?shard=0 as a query parameter.

Missing a label will be treated as mismatch.

This patch is not ident on purpose and will follow be an indetnation
patch.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>