cmake: Remove FindBoost.cmake
We used to include our own copy of FindBoost.cmake to allow using a
newer copy of boost than was supported by the installed cmake.

Our copy is getting old and we no longer provide our own copy of
boost, so it seems reasonable to assume that the system cmake supports
the system version of boost.

A quick check shows that Fedora 27 has cmake 3.11, which supports
boost versions < 1.68 and Fedora 28 and newer has 3.14 which support
boost versions < 1.71.

Fixes #691

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20191009210721.112710-1-espindola@scylladb.com>