stall_detector: Add a counter for stall detector report
It is helpful to know that a stall was detected so the user could go and
find it.

This patch adds a counter for the number of reported stalls.
Each time a stall is reported, the counter increases.

This will allow users to monitor, set alarms and annotations in the
graph to identify when a stall was detected.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>
Message-Id: <20191017145434.8630-1-amnon@scylladb.com>