net: posix: detemplatize posix_ap_server_socket_impl
Make the transport a member variable instead of a template parameter.

Since the per-transport connection queue now becomes unified,
the map key needs to include the transport (to disambiguate
an incoming TCP connection to 1.2.3.4:5 and an incoming SCTP
connection to 1.2.3.4:5. So the key becomes a
tuple<transport, socket_address>, and a hash function for that key
is defined.