noncopyable_function: do not copy uninitialized data
The direct path of noncopyable_function's move operation always
copies 32 bytes, of which the function instance can occupy a portion.
This triggers uninitialized variable warnings when the compiler
notices that it initialized fewer than 32 bytes but is copying more.

The warning does not indicate a bug, since this extra data will not
be used, but still breaks the build.

Fix by specializing the direct move function according to the
amount of data we initialized. In practice only four specializations
will be generated, for sizes 8, 16, 24, and 32, so this doesn't add
much bloat.
Message-Id: <20191021134030.15090-1-avi@scylladb.com>