inet_address: Make inet_address == operator ignore scope (again)
Fixes #704

Because of how inet_address:es are typically used in client apps (esp.
certain database apps), to achieve full transparent behaviour, we
should ignore the scope part of the address when doing comparison.

This is all based in apps doing serialization of addresses, and
comparing them on different hosts. Note that adding scopes to
serialization in these cases are not an options, since the scope
is not relevant on a different host, thus should be considered
serialization transient. However, same serialized addresses are
compared to "local" versions of the same, but with scope in them.
And those compare as different. We don't want that. So should make
comparison ignore scope, and treat this as wholly local/transient.
Yet propagate it so bind etc works.

This is all somewhat despicable behaviour, but consistent with the
almighty jdk forefather, where any address is considered equal iff
the bytes comprising it are same length and same content. And scope
is not part of the bytes.

Message-Id: <20191120155849.25206-1-calle@scylladb.com>